+++ CHANGE_LOG.md
@@ -1,10 +1,24 @@
-### 1.0.2 - 19/04/2022
+### 1.1.0 - 23/04/2022
 
-- Remove Henkel - they exited the RU market
+- Remove Henkel and AB InBev - they exited RU market
+- Add additional grade C, D brands
+- Add supermarkets in:
+
+  - Estonia (Coop, Rimi, Selver, Prisma Peremarket)
+  - Finland (K-Market, Alko, Prisma, S-Market, Sale, Alepa, Herkku, Sokos Herkku, Mestarin Herkku)
+  - Germany (Rewe, Amazon, Flaschenpost, Bringmeister, Lidl)
+  - Latvia (Barbora.lv, Rimi, Top!, Citro, Drogas)
+  - Lithuania (lastmile.lt, Čia Market, Eurokos, Vynoteka.lt, Rimi)
+  - Sweden (City Gross, Coop, ICA, Willys, Hemkop)
+  - Spain (Hipercor, El Corte Inglés, Caprabo, Leclerc)
+  - Ukraine (Мегамаркет, Zakaz, Auchan, Варус, Novus, Pchelka, ЕКО-маркет, Ultramarket, Cosmos, Tavriav, Fozzy, АТБ-Маркет, Сільпо, Наш Край)
+  - United States (Costco, Walmart, Target, Instacart, Amazon, Google Shopping, Kroger, Whole Foods, Harris Teeter, Safeway, Food Lion, Albertsons, Vons, Jewel-Osco, Acme, Stop & Shop, Giant Food Stores, Giant Food, Hannaford)
+
+- Overlay footer on top of cookies banners
+- Fix bug with close banner button
 
 ### 1.0.1 - 18/04/2022
 
 - Add Romania supermarkets
-- Add additional grade F brands 
+- Add additional grade F brands
 - Bug fixes
-